Musical Genre Classifier trained on the Marsyas dataset
Musical Genre Classifier trained on the Marsyas dataset. For Audio Feature extraction we are using MFCCs. MFCCs for training this version of the model are a subset of the entire dataset as the dataset is pretty large. There are 2 models, the first one uses a simple DNN, the second one makes use of a CNN. More information can be found in the readme